Fix issues outlined in StackOverflow issue http://stackoverflow.com/questions/26786246/how-to-parse-many-xml-files-with-cdata-in-a-folder-and-import-to-mysql-using-j/26788580
This commit does two things:

1.  It configures the ItemReader in the project to use the
MultiResourceItemReader so that it will read all xml files in the
classpath:/films directory.
2.  It configures the Film class to correctly parse the XML.  I added
the dependency for Eclipselink's Moxy project to provide more robust XML
parsing options.

This commit does not address parsing the actors since that is a list and
the domain object was only looking for a single actor.  However, that
should be simple to implement.